Use FDW to query multiple servers as shards (#320)
This commit mainly meets the needs of users to query multiple clusters
as external shards. fdw treats each data source as a whole without
knowing its internal structure. It keeps requesters and data sources
properly decoupled to maintain generality.

Add a new catalog table pg_foreign_table_seg to enable multiple shards
in foreign table. The foreign table should be treated as a shard table
with strewn locus. Each QE scanning the foreign should got a shard
from pg_foreign_table_seg.

Considering that the size of the computing cluster and the number of
shards of the foreign table may be inconsistent. Use flexible gang to
generate the same number of scan nodes as foreign table shards.

Considering that the data bandwidth between different data centers is
limited, we need to reduce the data transmission of fdw as much as
possible. Pushing the execution node down to the remote end as much
as possible can reduce data transmission.

If all tables of a subtree are distributed in the same foreign server
collection, It can be pushed down. But in mpp-fdw, we should consider
if a table only joinning the shared in same foreign server. So a new
system attribute gp_foreign_server was add to the foreign table. If
the customer add "t1.gp_foreign_server = t2.gp_foreign_server" to
join condition. It should be pushed down.

We can only push down the first stage of the two-stage aggregation.
Multi-stage aggregation will use some intermediate types. Some of
these intermediate types are external types that can be output
externally, such as count, min, max, and sum. The intermediate and
final types of these types are identical. Others are more complex
internal types, such as avg, whose intermediate type is inconsistent
with the final type and must be converted using a final function.
Since the local node in FDW serves as a standard client to exchange
data with the remote server, these internal types cannot be
transmitted. So some of the aggregate functions such as "avg"
should not be pushed down now.