[AQUMV] Check relation actually has children when view matching.
We arbitrarily set inh to false with a FIXME before this commit.
It should be corrected later during planning, but it's earlier
as we are doing view matching on parse tree.
Correct inh and skip it if has children.

Authored-by: Zhang Mingli avamingli@gmail.com
(cherry picked from commit 907985e1e9f524618a5429d70ceb25d1f9936279)