Fix:perfmon record wrong query history timestamp
When the qd node triggers the hook to send qlog messages to gpsmon, the current status
of the query is recorded in a local file. Subsequently, gpsmon forwards the qlog to gpmmon,
and gpmmon determines whether this query has been completed based on the status of the qlog
to decide whether to clear the query-related information in the memory. The reason for this
bug is that gpmmon does not judge based on the status of the qlog but directly judges through
the file that records the query status. When gpsmon has not yet forwarded the qlog information
to gpmmon, and gpmmon directly checks the status of the query in the file, it will accidentally 
clear the query information in the memory.