Remove redundant param from 'ivm_visible_in_prestate' call. . (#578)
At first glance, i did get the neediness of this parameter.  The visibility of tuple on QE does not depend on segment id, only on exported snapshot. So, this is clearly unneeded. But... why IVM test are not failing on this? This got me some more digging in CBDB internals, I actually think all this code is simply dead code for Cloudberry. 

Prestate visibility in IVM is something that matters when multiple relation are modified via SINGLE statement. I don't think we can even do this in Cloudberry:
1)  Multiple write CTE are not possible, because  there could be only one writer gang. Ref: https://github.com/cloudberrydb/cloudberrydb/commit/bfcb78829d50e31a49ce18d36d9c3af9b8a2ffe3
2) After statement are also prohibited, I get 'Triggers for statements are not yet supported.' error with them. 

Are there any other way for multiple relation modification in PostgreSQL that works for GP/CBDB? 
And, yes, even when this will (if even) be supported, gpseg id is anyway redundant.