Datalake: change datalake uriparser function name
The uriparser function names are the same with pxf extension, which will cause
an error when creating a pxf external table:

"cannot execute pxfprotocol_validate_urls outside protocol manager
(pxfprotocol.c:57)"

The reason is that the pxf routine will invoke the function in
datalake(eg. parseGPHDUri),
and this is not expected.