Fix explain analyze printing "never executed" uncorrectly (#672)
explain analyze insert into test select i from
generate_series(0,100) as i;

Insert on test (never executed)
    ->  Result (actual time=0.021..0.046 rows=38 loops=1)
	    ->  Result (actual time=0.021..0.041 rows=38 loops=1)
		    ->  Function Scan on generate_series (actual
			    time=0.018..0.026 rows=101 loops=1)

As shown above, insert node is considered never executed. Actually,
it has been executed. This is due to the collected nloops of that node
is wrong. When the nloops in the instrument is 0, the plan node will
be considered as 'never executed'.

In some plan nodes, the rows returned is 0, but the nloops is not,
fix related code in gp_explain.c

(cherry picked from commit 87971ba4aa9236ea32fae281d0338abe491114c7)