Fix parallel scan for pax tables
Parallel worker processes are assigned a set of micro partitions
to scan. All parallel workers should arrive and wait before the next
step. The worker process will run:
1. Attach to the build barrier
2. Scan the inner table to build the shared hash table
2.1  Request a micro partition file to scan
2.2  Scan tuples to build hash table
3. Arrive and wait for all build participants
4. At this moment, the inner table has been fully scanned to build
   hash table. There may be some worker processes that still not
   reach the phase of building hash table.
5. Run probing phase to produce output tuples.

The problem is that the worker process for scanning pax table acquires
micro partition files before attaching the build barrier of parallel
hash join. When the other worker process finishes the phase of building
hash table, it sees no other participants are building the hash table,
and no tuple ranges are left. So, the running phase of parallel hashjoin
moves on. When the worker process that acquires micro partition file
and not enters the phase of building hash table, will have no chance to
add its tuples to the hash tables.
The fix is simple that lazily acquires the micro partition files when
tuples are requested.