[perfmon] Fix some issues in gpmmon
- write the time with timezone
- set the default value of tsubmit/tstart/tfinish to -infinity
- modify agg_dup
    agg_dup is used to copy the unfinished query into a new hash table
and clean the old one. It reads the query text file and gets status
from it. From the status, it can know if the query is finished. But
the queries in the hash table is a little behand the status from the
query text file , using the status to filter query may lead some issues
: finished query is not recorded or the tfinish time is empty. So we use
the status from the hash table to check the query.
    For query is recored or executed time less than the config time,
 just clean it.