[perfmon] Let gpmmon not to kill gpsmon when timeout getting response
Gpmmon polls the query info from gpsmon each 'quantum'
time. But when this is no query running, gpsmon doesn't
response to it. Then gpmmon considers gpsmon dead
and restarts it. This will lead the whole system cannot
run correctly when busy restarting the gpsmon.

Fix it by letting the gpmmon ignore the timeout event
when waiting the respose from gpsmon.