vectorization: support whitelist
The original implementation does not use oid to determine whether fallback or not, but uses oid to obtain the function name and judge by the function name, which leads to many operations that should not be vectorized, such as cash + cash. In this mr, oids are used to determine whether vectorization is supported. Oids that support vectorization need to provide build functions and extra check functions