Fix the proc cpu percent got by gpsmon
The proc cpu percent got by gpsmon sometimes is 0 although
the proc cpu cost is not 0. The bug is caused by the
pidtab hash table's key. The key is not pid, but defined
as queryid + pid. In the case a session runs a bunch of
queries, there will be multiple entries in the pidtable
which have the same pid, and the cpu percent will be computed
multiple times for a proc. The cpu percent is computed as
nowcpucost-lastcpucost/timemow-lasttime, then the 
nowcpucost-lastcpucost is 0 if it is collected within
a minimal timeframe. Fix that by using pid as the hash
table's key.

By the way, did a small refactor to make the code more
clean.