Fix crash that syslog receives messages from threads
Syslog is designed by postgres upstream that targets process model.
Cloudberry has threads for some process, like receive thread in
interconnect. Different threads in the same process may write log
messages concurrently.

The syslog is a multi-producer-one-consumer model. The process/thread
is the producer that writes the log message to a pipe. The auxiliary
process of postgres kernel, syslogger, is the consumer that receives
log messages from the pipe.

If the log message is too large, the message will be splited to multiple
chunks. Each chunk message has a PipeProtoHeader, with chunk_no increased
from 0. For example, a message is splited to 3 chunks:

       +----------------+---------------+
       |PipeProtoHeader |               |
       | pid,tid,len    |   payload0    |
       |  chunk_no=0    |               |
       |  is_last='f'   |               |
       +----------------+---------------+

       +----------------+---------------+
       |PipeProtoHeader |               |
       | pid,tid,len    |   payload1    |
       |  chunk_no=1    |               |
       |  is_last='f'   |               |
       +----------------+---------------+

       +----------------+---------------+
       |PipeProtoHeader |               |
       | pid,tid,len    |   payload2    |
       |  chunk_no=2    |               |
       |  is_last='t'   |               |
       +----------------+---------------+

The logical message should be
       +----------------+---------------+
       |PipeProtoHeader |   payload0    |
       | pid,tid,len    |   payload1    |
       |  chunk_no=0    |   payload2    |
       |  is_last='t'   |               |
       +----------------+---------------+
if the buffer is large enough to put all payload into one chunk.

The different threads in the same process share the same pid. They should
have their own queue to assemble chunks to a complete payload, or
the payload boundary is broken. The payload0 has additional meta fields
that represent messages like time, gp_session_id, etc. If the payload
boundary is wrong, the broken meta fields will get corrupted time.
The coredump happens in the next format functions.