SINGLENODE and EntryDB materialized view status maintenance optimization.
In SINGLENODE mode or when operating as an entry db, we cannot use
the extend libpq to send messages because we are already functioning in
a QD role. Processing modified relations at `EndModifyTable()` is
too late, especially since materialized views are updated when
the executor ends.

This change moves the processing of modified relations to
`ExecModifyTable()` to ensure timely handling of modifications. This
avoids issues with materialized view updates and ensures correct
behavior in SINGLENODE or entry DB scenarios.

Authored-by: Zhang Mingli avamingli@gmail.com