Fix the invalid cast from GArrowXXXScalar to GArrowNullScalar
The same base class is used for GArrowNullScalar and GArrowXXXScalar
(such as GArrowInt32Scalar, ...). Therefore, it is incorrect to represent an
invalid scalar (which means null in the database) using GArrowNullScalar by
calling garrow_null_scalar_new(). Instead, we should create a scalar which
is_valid is set to false using garrow_invalid_scalar_new().