Change directory table directory's file path from relfilenode to relid.
At first, we use relfilenode to construct the directory table's
directory path. However, relfilenode may be different between QD and
QE, this will cause the directory table's file paths are different which
will be confused in using. As a result, we will use relid that is same
between QD and QE to construct the directory table's file path in this
commit.

Authored-by: Zhang Wenchao zhangwenchao@apache.org