Statistic: Add STATISTIC_KIND_NDV_BY_SEGMENTS used to calculate ndistinct on segment
When there are aggregation statements in SQL, the optimizer usually generates
a logical GbAgg operator. For this type of logical operator, if distinct does
not exist, the physical operator can be one-stage or two-stage.

The cost of two-stage aggregation will be calculated by  NDV(number of distinct
values). But the problem is that the statistics in CBDB are "global". It means
that the global NDV value may small, but the NDV value on segments can be larger.

Therefore, the current commit introduces a new statistical kind which named
`STATISTIC_KIND_NDV_BY_SEGMENTS`, this kind of statistical will calculate the
NDV on the segment and accumulate each NDV of singlesegment on the master.

(cherry picked from commit 9d7820e43dcb2ea9f9b0301697d49099411e94a4)