Fix different aggsplit in the final agg stage
ORCA will generate a plan which will have not the same aggsplit
in the final agg stage. In fact, the strange thing is that ORCA
has not changed the logic of aggsplit in recent versions. So I
am not sure  why vectorization didn’t have this problem before.
I suspect it might be caused by different plans.

Anyway, According to the existing logic, the aggsplit in the
executor context is only used to verify that the aggsplits of
different aggrefs all same, which is unnecessary.