Fix invalid attribute mapping when runtime filter pushdown is enabled.
Seqscan is a special case, it's targetlist is a projection of the
relation's attributes, so we need to find the attribute number of
the column in the relation.
While TargetEntry::resno is just the attribute number for dynamic
seqscan.

(cherry picked from commit 13a3dc4dfc6b8a3c730f3c8b3fed837df473c177)