Fix: quick exit in dumptuples if query execution is finished
If QueryFinishPending is set when query is running into dumptuples, the
tuplecontext is reset but memtuples are not cumsumed. When query is running
into dumptuples again, tuplesort_sort_memtuples will access these memtuples,
and the memory allocated in tuplecontext is already freed, this will cause
invalid memory access.

To avoid this situation, do nothing in dumptuples if QueryFinishPending is
set.