Fix flaky case resource_queue_multi_portal
In isolation2 resource_queue_multi_portal testcase,
we use pg_cancel_backend to terminate query.
The pg_cancel_backend function only sends SIGINT information
to the backend, and the backend process may not have time to
process this signal.The rsqcountvalue will not decrease.
So we wait for 1 second after canceling the query,
basically ensuring that the query is terminated.