Backport: Compute aggregate argument types correctly in transformAggregateCall(). (#1500)
This pr fixes https://www.postgresql.org/support/security/CVE-2023-5868 in cloudberry

https://git.postgresql.org/cgit/postgresql.git/commit/?id=3b0776fde56763c549df35ce9750f3399bc710b2

===

transformAggregateCall() captures the datatypes of the aggregate's arguments immediately to construct the Aggref.aggargtypes list. This seems reasonable because the arguments have already been transformed --- but there is an edge case where they haven't been. Specifically, if we have an unknown-type literal in an ANY argument position, nothing will have been done with it earlier.  But if we also have DISTINCT, then addTargetToGroupList() converts the literal to "text" type, resulting in the aggargtypes list not matching the actual runtime type of the argument.  The end result is that the aggregate tries to interpret a "text" value as being of type "unknown", that is a zero-terminated C string.  If the text value contains no zero bytes, this could result in disclosure of server memory following the text literal value.

To fix, move the collection of the aggargtypes list to the end of transformAggregateCall(), after DISTINCT has been handled. This requires slightly more code, but not a great deal.

Our thanks to Jingzhou Fu for reporting this problem.

Security: CVE-2023-5868
(cherry picked from commit f35b3326d1ff453e152a65aa5813107c448b3113)